refactor: `VisitEntity`를 `VisitDto`로 이름 변경 `VisitEntity` 클래스의 이름을 `VisitDto`로 변경하여, 해당 클래스가 데이터 전송 객체(DTO)임을 더 명확히 나타냅니다. 이 변경 사항에는 다음이 포함됩니다: 	•	클래스 파일명을 `VisitEntity.java`에서 `VisitDto.java`로 변경 	•	`VisitService`, `VisitJpaRepository`, `VisitRepository` 등 프로젝트 전체에서 `VisitEntity` 사용 부분을 모두 `VisitDto`로 변경 	•	관련 패키지 import 구문도 모두 수정
