+++ src/main/java/com/realyoungk/sdi/repository/VisitJpaRepository.java
@@ -1,7 +1,7 @@
 package com.realyoungk.sdi.repository;
 
-import com.realyoungk.sdi.entity.VisitEntity;
+import com.realyoungk.sdi.dto.VisitDto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VisitJpaRepository extends JpaRepository<VisitEntity, String>, VisitRepository {
+public interface VisitJpaRepository extends JpaRepository<VisitDto, String>, VisitRepository {
 }
